Simplified butterworth() function
A tricky memory allocation scheme was used, and was quite ineffective.
The purpose was to avoid a huge memory allocation, by allocating output
arrays lines one at a time. However, it all ended with a call to
np.asarray(), which did a huge memory allocation, *and* copy everything.

Replaced by a simpler scheme, that turns out to be faster as well.