Changed output of region signal extracting functions
img_to_signals_labels() and img_to_signals_maps() now always return a
number of signals equal to the number of regions passed as parameter. In
the previous state, only regions not fully outside the mask were
considered, the number of output signals thus could vary. This is not
the case anymore.

Tests have been adapted.