[TST] Fix failing test
A test is failing on Travis (np.diff(objective) <= 0 in
test_group_sparse_covariance.py:51), that does not occur of the
developer's machine. np.diff(objective) may be hitting the numerical
noise floor (1e-14 or so). Some values may then be slightly positive.
This is most probably due to a different numerical library which leads to
a slightly different result.

The fix consists in decreasing the number of iterations so that optimization
stops before hitting the noise floor.