NF: Add Massively Univariate Linear Model with permuted OLS (v2).
This PR provides massively univariate linear models estimation using
OLS regression and permutation testing. The code is designed to
complete a very large number of permutations (> 100000) on a large
problem (thousands of targets variables, namely the brain voxels)
within a few minutes.
As compared to the v1 PR, we do not split the targets so as to perform
their analysis on parallel computing units, but we delegate a fraction
of permutations to the parallel computing units. This solves the
problem we had ensuring the data are permuted in the same manner by
each worker. The code is also simpler as the max F-score across data
descriptors can be computed directly after each permutation.
We assume that we perform analysis of a reasonable number of tested
variates and that each computing unit can process the target variates
without the need to split it. Typically, neuroimaging-genetic problems
(and particularly Genom-Wide Association Studies) require specific
code optimizations that are not straightwforward with the new design
proposed in this PR.