Get coveralls report back
The reason coveralls reports did not have any data was because .coverage
was generated in /tmp but coveralls was running from the git
checkout.

Now using TEST_RUN_FOLDER for where the tests are run from and copying
.coverage to the git checkout folder.

Note coveralls need to run from the git checkout folder because it
relies on git commands, for example to retrieve the git commit id.