Simplify matplotlib backend choice
and test skipping when matplotlib is not installed.

Matplotlib backend is set once and for all in
nilearn/plotting/__init__.py. Agg is used only when DISPLAY is unset.

Failing to import matplotlib is only ok when running the
tests. nose.SkipTest are not needed anymore in nilearn.plotting
submodules.