ENH: faster faasta (for TV-l1)
Careful timing / profiling shows that the dual-gap of the proximal
operator is sometimes decreased too far. As a result, the proximal
operator spends a lot of time trying to reach this dual gap, for no good
increase in overall energy. Profiling the code shows that this is the
number one loss of time.

Given that with warm-restart well implemented, decreasing on the fly the
tolerance of the proximal isn't very costly, we can rely fully on the
adaptative strategy, rather than on the paper of Schmidt et al (ie
implement FAASTA as in the paper). I have done this here. Printing values
of energies on the Haxby dataset shows that energy is decreased as well.

The code seems to be around 1.5 times faster.

I am sending a PR to see how this behaves on Travis / Circle