DOC: rewrite the parcellating section
The goal of this rewrite is to move all the code sample to the
notebook-like example, and keep only a high-level discussion in the
narrative documentation.

Indeed, modifications of the code in the examples (as in the previous
commit) cause code-rot in the narrative documentation that is very
hard to control.