DOC + ENH: tutorial example and get_data_dirs func
The get_data_dirs function is useful to debug during a tutorial while a
user is not putting the data in the right directory when copying from a
pen drive (this happens often).

The example should probably be the 2nd one in a beginner tutorial, to
make people a bit more familiar with input types.