Also allow scipy.nmimage.measurements.sum as a reducting function
Forgot to add sum as a possible reducing function in the previous commits.
Technically we could check the signature and output types
of all functions in scipy.nmimage.measurements to check which ones
should be allowed. Or generalise the code behind signal_extraction
to accept any user function that follows a certain signature.