[FIX] LGTM.com recommendations (#2980)
* [RIX] LGTM.com recommendation: Except block handles 'BaseException'

Except block directly handles BaseException.

https://www.flake8rules.com/rules/E722.html

A bare except: clause will catch SystemExit and KeyboardInterrupt
exceptions, making it harder to interrupt a program with Control-C,
and can disguise other problems. If you want to catch all exceptions
that signal program errors, use except Exception: (bare except is
equivalent to except BaseException:).

* [FIX] LGTM.com recommendation: Module is imported more than once

This import of module os is redundant, as it was previously imported on line 44.

* [FIX] LGTM.com recommendation: Unused import

Import of 'Axes3D' is not used.
Import of 'sklearn' is not used.